# apps/api/spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Convixx API",
    "description": "Multi-tenant AI Voice Calling Platform backend. RAG-based question answering with self-hosted LLM and OpenAI fallback.\n\n**Documentation:** Interactive UI at `/docs`. Raw OpenAPI JSON at `/docs/json` (same origin as this API).",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host (recommended for Try it out)"
    },
    {
      "url": "https://convixx.in",
      "description": "Production"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "AdminToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-token",
        "description": "Admin token for customer APIs. Set in .env as ADMIN_TOKEN."
      },
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Customer API key. Scopes KB, agents, ask, chat, voice to that customer."
      }
    },
    "schemas": {
...
      "CompanyPhoneNumber": {
        "type": "object",
        "description": "Row in `company_phone_numbers`. One company can own many numbers; `customer_id` is unique per `phone_number`.",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "customer_id": {
            "type": "string",
            "format": "uuid"
          },
          "phone_number": {
            "type": "string",
            "example": "+9102048555864"
          },
          "label": {
            "type": "string",
            "nullable": true,
            "example": "Sales line"
          },
          "default_agent_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Agent this number routes inbound calls to. Null falls back to the company's oldest active agent."
          },
          "is_primary": {
            "type": "boolean",
            "description": "At most one primary number per customer; used as the default outbound caller ID."
          },
          "is_enabled": {
            "type": "boolean"
          },
          "provider_override_id": {
            "type": "string",
            "nullable": true,
            "description": "Reserved for future carrier-migration support (§8.5 roadmap). Not currently settable via this API."
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CompanyPhoneNumberCreate": {
        "type": "object",
        "required": [
          "phone_number"
        ],
        "properties": {
          "phone_number": {
            "type": "string",
            "minLength": 3,
            "example": "+9102048555864"
          },
          "label": {
            "type": "string",
            "minLength": 1
          },
          "default_agent_id": {
            "type": "string",
            "format": "uuid"
          },
          "is_primary": {
            "type": "boolean",
            "description": "If true, unsets any other primary number for this customer first."
          },
          "is_enabled": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "CompanyPhoneNumberUpdate": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "minLength": 1,
            "nullable": true
          },
          "default_agent_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "is_enabled": {
            "type": "boolean"
          },
          "is_primary": {
            "type": "boolean",
            "description": "If true, unsets any other primary number for this customer first."
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/customers/{customerId}/phone-numbers": {
      "get": {
        "tags": [
          "Company Phone Numbers"
        ],
        "summary": "List phone numbers",
        "description": "Returns every number owned by this company, primary first. Requires x-admin-token.",
        "security": [
          {
            "AdminToken": []
          }
        ],
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompanyPhoneNumber"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Company Phone Numbers"
        ],
        "summary": "Add a phone number",
        "description": "Adds a number to this company. Fails with 409 if the number is already assigned to any company (numbers are globally unique). Requires x-admin-token.",
        "security": [
          {
            "AdminToken": []
          }
        ],
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyPhoneNumberCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyPhoneNumber"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "default_agent_id does not belong to this customer"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Customer not found"
                }
              }
            }
          },
          "409": {
            "description": "Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "This phone number is already assigned to a company"
                }
              }
            }
          }
        }
      }
    },
    "/customers/{customerId}/phone-numbers/{id}": {
      "put": {
        "tags": [
          "Company Phone Numbers"
        ],
        "summary": "Update a phone number",
        "description": "Partial update. Setting `is_primary: true` unsets any other primary number for this customer first. Requires x-admin-token.",
        "security": [
          {
            "AdminToken": []
          }
        ],
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyPhoneNumberUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyPhoneNumber"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "default_agent_id does not belong to this customer"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Phone number not found for this customer"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Company Phone Numbers"
        ],
        "summary": "Delete a phone number",
        "description": "Refuses (409) to delete the primary number while other numbers exist for this customer — set another number as primary first. Requires x-admin-token.",
        "security": [
          {
            "AdminToken": []
          }
        ],
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "example": {
                  "message": "Phone number deleted"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Phone number not found for this customer"
                }
              }
            }
          },
          "409": {
            "description": "Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Cannot delete the primary number while other numbers exist — set another number as primary first"
                }
              }
            }
          }
        }
      }
    },
...
  }
}